✨ 응답 메소드 클래스 내의 메소드로 이동
